@@ -1,236 +1,284 @@
-﻿namespace BrowserSearch
+﻿namespace BrowserSearch;
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text.RegularExpressions;
+using System.Threading;
+using CommandLine;
+
+public class Program
 {
-	using CommandLine;
-	using System;
-	using System.Collections.Generic;
-	using System.Diagnostics;
-	using System.IO;
-	using System.Linq;
-	using System.Runtime.InteropServices;
-	using System.Threading;
+    private const int CommandCompleteWaitDefaultMs = 5000;
+    private string[] startingPhrases =
+    {
+            "what is the",
+            "how to",
+            "what color",
+            "describe the",
+            "where is it",
+            "why is it",
+            "show tall",
+            "the length of",
+            "who is",
+            "who did",
+            "where is",
+            "are the",
+            "the best of",
+            "which of",
+            "the location of",
+        };
 
-	public class Program
-	{
-		public static int Main(string[] args)
-		{
-			int exit = 0;
-			ParserResult<CommandLineOptions> result = Parser.Default.ParseArguments<CommandLineOptions>(args)
-			.WithParsed(options =>  // options is an instance of Options type
-			{
-				exit = new Program().Run(options);
-			})
-			.WithNotParsed(errors =>  // errors is a sequence of type IEnumerable<Error>
-			{
-				exit = 0;
-			});
+    public static int Main(string[] args)
+    {
+        int exit = 0;
+        ParserResult<CommandLineOptions> result = Parser.Default.ParseArguments<CommandLineOptions>(args)
+        .WithParsed(options =>  // options is an instance of Options type
+        {
+            exit = new Program().Run(options);
+        })
+        .WithNotParsed(errors =>  // errors is a sequence of type IEnumerable<Error>
+        {
+            exit = 0;
+        });
 
-			return exit;
-		}
+        return exit;
+    }
 
-		private readonly IList<string> availableSearchWords;
-		private OSPlatform? osPlatform = null;
+    private readonly IList<string> availableSearchWords;
+    private OSPlatform? osPlatform = null;
 
-		public Program()
-		{
-			this.availableSearchWords = LoadAvailableWords();
-			this.DetermineOSPlatform();
-		}
+    public Program()
+    {
+        this.availableSearchWords = LoadAvailableWords();
+        this.DetermineOSPlatform();
+    }
 
-		private int Run(CommandLineOptions options)
-		{
-			try
-			{
-				string browser = options.BrowserName;
-				if (string.IsNullOrWhiteSpace(options.BrowserName))
-				{
-					browser = "OS default";
-				}
-				Console.WriteLine($"Using browser: {browser}");
+    private int Run(CommandLineOptions options)
+    {
+        for (int i = 0; i < options.CycleCount; i++)
+        {
+            Console.WriteLine($"--------------------- Executing cycle #{i + 1}/{options.CycleCount} ---------------------");
 
-				var wordHistory = new HashSet<string>();
-				var rnd = new Random();
+            // delay if we're repeating
+            if (i > 0)
+            {
+                // convert pause ms to minutes
+                double pauseMins = options.CyclePauseMs / 60000.0;
+                Console.WriteLine($"Pausing for {Math.Round(pauseMins, 2)} minute(s)");
+                Thread.Sleep(options.CyclePauseMs);
+            }
 
-				ExecuteCommandBefore(options);
+            // execute an entire cycle
+            this.ExecuteCycle(options);
+        }
 
-				for (int i = 0; i < options.SearchCount; ++i)
-				{
-					string words = null;
-					do
-					{
-						words = string.Join("+", this.GetWords(rnd, options).Select(w => w.ToLower().Trim()).OrderBy(w => w));
+        return 1;
+    }
 
-					} while (wordHistory.Contains(words));
+    private void ExecuteCycle(CommandLineOptions options)
+    {
+        try
+        {
+            string browser = options.BrowserName;
+            if (string.IsNullOrWhiteSpace(options.BrowserName))
+            {
+                browser = "OS default";
+            }
+            Console.WriteLine($"Using browser: {browser}");
 
-					Console.WriteLine($"Opening search #{i + 1} using \"{words.Replace("+", " ")}\"");
-					wordHistory.Add(words);
+            var wordHistory = new HashSet<string>();
+            var random = new Random();
 
-					// search bing
-					string url = FormatSearchEngineUrl(words);
-					this.OpenBrowser(url, options);
+            ExecuteCommandBefore(options);
 
-					// sleep if not the last time
-					if (i + 1 < options.SearchCount)
-					{
-						Thread.Sleep(options.SearchDelay + rnd.Next(1000)); // add a random delta to help look non-automated
-					}
-				}
+            for (int i = 0; i < options.SearchCount; ++i)
+            {
+                string searchPhrase = null;
+                do
+                {
+                    searchPhrase = string.Join("+", this.GetWords(random, options).Select(w => w.ToLower().Trim()));
 
-				this.ExecuteCommandAfter(options);
-			}
-			catch (Exception)
-			{
-				throw;
-			}
+                } while (wordHistory.Contains(searchPhrase));
 
-			return 1;
-		}
+                Console.WriteLine($"Opening search {i + 1}/{options.SearchCount} using \"{searchPhrase.Replace("+", " ")}\"");
+                wordHistory.Add(searchPhrase);
 
-		private static string FormatSearchEngineUrl(string words)
-		{
-			return $"https://www.bing.com/search?q={words}";
-		}
+                // search bing
+                string url = FormatSearchEngineUrl(searchPhrase, random);
+                this.OpenBrowser(url, options);
 
-		private void OpenBrowser(string url, CommandLineOptions options)
-		{
-			// hack because of this: https://github.com/dotnet/corefx/issues/10361
-			if (this.osPlatform == OSPlatform.Windows)
-			{
-				url = url.Replace("&", "^&");
-				_ = Process.Start(new ProcessStartInfo("cmd", $"/c start {options.BrowserName} \"{url}\"") { CreateNoWindow = true });
-			}
-			else if (this.osPlatform == OSPlatform.Linux)
-			{
-				_ = Process.Start("xdg-open", url);
-			}
-			else if (this.osPlatform == OSPlatform.OSX)
-			{
-				_ = Process.Start("open", url);
-			}
-			else
-			{
-				throw new ApplicationException($"Internal error; invalid OS defined in {nameof(osPlatform)}");
-			}
-		}
+                // sleep if not the last time
+                if (options.SearchDelayMs > 0 && i + 1 < options.SearchCount)
+                {
+                    Thread.Sleep(options.SearchDelayMs + random.Next(500, 1001)); // add a random delta to help look non-automated
+                }
+            }
 
-		private static void ExecuteCommandBefore(CommandLineOptions options)
-		{
-			if (string.IsNullOrWhiteSpace(options.CommandBefore))
-			{
-				return;
-			}
+            ExecuteCommandAfter(options);
+        }
+        catch (Exception)
+        {
+            throw;
+        }
+    }
 
-			Console.WriteLine($"Executing command before search, \"{options.CommandBefore}\"");
-			ExecuteOperatingSystemCommand(options.CommandBefore, options.CommandBeforeWait);
+    private static string FormatSearchEngineUrl(string words, Random random)
+    {
+        return $"https://www.bing.com/search?q={words}&form=QBLH";
+    }
 
-			if (options.CommandBeforePause > 0)
-			{
-				Console.WriteLine($"Pause after executing command, before search, {options.CommandBeforePause} ms");
-				Thread.Sleep(options.CommandBeforePause);
-			}
-		}
+    private void OpenBrowser(string url, CommandLineOptions options)
+    {
+        // hack because of this: https://github.com/dotnet/corefx/issues/10361
+        if (this.osPlatform == OSPlatform.Windows)
+        {
+            // commented this out 10/24/2023; not sure why this replace was being done, but seems
+            // to work without it (for now)
+            //url = url.Replace("&", "^&");
 
-		private void ExecuteCommandAfter(CommandLineOptions options)
-		{
-			if (string.IsNullOrWhiteSpace(options.CommandAfter))
-			{
-				return;
-			}
+            _ = Process.Start(new ProcessStartInfo("cmd", $"/c start {options.BrowserName} \"{url}\"") { CreateNoWindow = true });
+        }
+        else if (this.osPlatform == OSPlatform.Linux)
+        {
+            _ = Process.Start("xdg-open", url);
+        }
+        else if (this.osPlatform == OSPlatform.OSX)
+        {
+            _ = Process.Start("open", url);
+        }
+        else
+        {
+            throw new ApplicationException($"Internal error; invalid OS defined in {nameof(osPlatform)}");
+        }
+    }
 
-			if (options.CommandAfterPause > 0)
-			{
-				Console.WriteLine($"Pause before executing command, after search {options.CommandAfterPause} ms");
-				Thread.Sleep(options.CommandAfterPause);
-			}
+    private static void ExecuteCommandBefore(CommandLineOptions options)
+    {
+        if (string.IsNullOrWhiteSpace(options.CommandBefore))
+        {
+            return;
+        }
 
-			Console.WriteLine($"Executing the command after search, \"{options.CommandAfter}\"");
-			ExecuteOperatingSystemCommand(options.CommandAfter, options.CommandAfterWait);
-		}
+        Console.WriteLine($"Executing command before search, \"{options.CommandBefore}\"");
+        ExecuteOperatingSystemCommand(options.CommandBefore, options.CommandBeforeWaitMs);
 
-		private static void ExecuteOperatingSystemCommand(string command, int waitToCompleteMs)
-		{
-			// a delimiter to separate the process command from the command arguments
-			const string separator = "::";
+        if (options.CommandBeforePauseMs > 0)
+        {
+            Console.WriteLine($"Pause after executing command, before search, {options.CommandBeforePauseMs} ms");
+            Thread.Sleep(options.CommandBeforePauseMs);
+        }
+    }
 
-			string args = null;
-			int separatorIndex = command.IndexOf(separator);
-			if (separatorIndex != -1)
-			{
-				// set args first, before changing command
-				args = command[(separatorIndex + separator.Length)..].Trim();
-				command = command.Substring(0, separatorIndex).Trim();
-			}
+    private static void ExecuteCommandAfter(CommandLineOptions options)
+    {
+        if (string.IsNullOrWhiteSpace(options.CommandAfter))
+        {
+            return;
+        }
 
-			try
-			{
-				Process process = null;
-				if (args is null)
-				{
-					process = Process.Start(command);
-				}
-				else
-				{
-					process = Process.Start(command, args);
-				}
+        if (options.CommandAfterPauseMs > 0)
+        {
+            Console.WriteLine($"Pause before executing command, after search {options.CommandAfterPauseMs} ms");
+            Thread.Sleep(options.CommandAfterPauseMs);
+        }
 
-				process.WaitForExit(waitToCompleteMs > 0 ? waitToCompleteMs : 5000); // default of 5 seconds
-			}
-			catch (Exception ex)
-			{
-				Console.Error.WriteLine($"Error excuting command \"{command}\". Reason: {ex.Message}");
-			}
-		}
+        Console.WriteLine($"Executing the command after search, \"{options.CommandAfter}\"");
+        ExecuteOperatingSystemCommand(options.CommandAfter, options.CommandAfterWaitMs);
+    }
 
-		private string[] GetWords(Random rnd, CommandLineOptions options)
-		{
-			var words = new HashSet<string>();
-			int count = rnd.Next(options.MinSearchWordCount, options.MaxSearchWordCount + 1); // max is exclusive so add 1
-			for (int i = 0; i < count; ++i)
-			{
-				string word;
-				do
-				{
-					word = availableSearchWords[rnd.Next(availableSearchWords.Count)];
-				} while (words.Contains(word));
+    private static void ExecuteOperatingSystemCommand(string command, int waitToCompleteMs)
+    {
+        // a delimiter to separate the process command from the command arguments
+        const string separator = "::";
 
-				words.Add(word);
-			}
+        string args = null;
+        int separatorIndex = command.IndexOf(separator);
+        if (separatorIndex != -1)
+        {
+            // set args first, before changing command
+            args = command[(separatorIndex + separator.Length)..].Trim();
+            command = command[..separatorIndex].Trim();
+        }
 
-			return words.ToArray();
-		}
+        try
+        {
+            Process process = null;
+            if (args is null)
+            {
+                process = Process.Start(command);
+            }
+            else
+            {
+                process = Process.Start(command, args);
+            }
 
-		private static IList<string> LoadAvailableWords()
-		{
-			const string FileName = "words.txt";
-			if (!File.Exists(FileName))
-			{
-				throw new FileNotFoundException($"Missing search words file, {FileName}");
-			}
+            process.WaitForExit(waitToCompleteMs > 0 ? waitToCompleteMs : CommandCompleteWaitDefaultMs);
+        }
+        catch (Exception ex)
+        {
+            Console.Error.WriteLine($"Error excuting command \"{command}\". Reason: {ex.Message}");
+        }
+    }
 
-			// read all non-blank lines, ignore those marked as a comment
-			return Array.AsReadOnly<string>(File.ReadAllLines(FileName)
-				.Where(l => !l.StartsWith("#") && l.Trim().Length > 0)
-				.ToArray());
-		}
+    private string[] GetWords(Random random, CommandLineOptions options)
+    {
+        var words = new HashSet<string>();
+        int count = random.Next(options.MinSearchWordCount, options.MaxSearchWordCount + 1); // max is exclusive so add 1
+        for (int i = 0; i < count; ++i)
+        {
+            string word;
+            do
+            {
+                word = availableSearchWords[random.Next(availableSearchWords.Count)];
+            } while (words.Contains(word));
 
-		private void DetermineOSPlatform()
-		{
+            words.Add(word);
+        }
 
-			if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-			{
-				this.osPlatform = OSPlatform.Windows;
-			}
-			else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-			{
-				this.osPlatform = OSPlatform.Linux;
-			}
-			else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-			{
-				this.osPlatform = OSPlatform.OSX;
-			}
-			else
-			{
-				throw new ApplicationException($"Unsupported platform to launch process, {RuntimeInformation.OSDescription}");
-			}
-		}
-	}
+        // now insert a starting phrase
+        var listWords = words.ToList();
+        string startingPhrase = startingPhrases[random.Next(startingPhrases.Length)];
+        listWords.Insert(0, startingPhrase);
+
+        return [.. listWords];
+    }
+
+    private static IList<string> LoadAvailableWords()
+    {
+        const string FileName = "words.txt";
+        if (!File.Exists(FileName))
+        {
+            throw new FileNotFoundException($"Missing search words file, {FileName}");
+        }
+
+        // read all non-blank lines, ignore those marked as a comment
+        return Array.AsReadOnly<string>(File.ReadAllLines(FileName)
+            .Where(l => !l.StartsWith("#") && l.Trim().Length > 0)
+            .ToArray());
+    }
+
+    private void DetermineOSPlatform()
+    {
+
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+        {
+            this.osPlatform = OSPlatform.Windows;
+        }
+        else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+        {
+            this.osPlatform = OSPlatform.Linux;
+        }
+        else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+        {
+            this.osPlatform = OSPlatform.OSX;
+        }
+        else
+        {
+            throw new ApplicationException($"Unsupported platform to launch process, {RuntimeInformation.OSDescription}");
+        }
+    }
 }
